move external clients to lib/llm
