+++ evals/llm_clients/hn_customOpenAI.ts
@@ -1,6 +1,6 @@
 import { EvalFunction } from "@/types/evals";
 import { z } from "zod";
-import { CustomOpenAIClient } from "@/examples/external_clients/customOpenAI";
+import { CustomOpenAIClient } from "@/lib/llm/customOpenAI";
 import OpenAI from "openai";
 import { Stagehand } from "@/dist";
