+++ evals/llm_clients/hn_langchain.ts
@@ -1,6 +1,6 @@
 import { EvalFunction } from "@/types/evals";
 import { z } from "zod";
-import { LangchainClient } from "@/examples/external_clients/langchain";
+import { LangchainClient } from "@/lib/llm/langchain";
 import { ChatOpenAI } from "@langchain/openai";
 import { Stagehand } from "@/dist";
